requirements.txt
matplotlib==3.9.0
numpy==2.0.0
Pillow==10.3.0
pytorch3d==0.7.4
torch==2.3.1
torch_directml==0.2.2.dev240614
torchvision==0.18.1
tqdm==4.66.4
